Updating .csproj file and removing dependency on log4net
Now multitargeting to allow FryProxy to be used with any version of
.NET Framework from 4.5 forward, and with .NET Core. Also updating to
the latest project format.

Additionally, this commit removes the dependency on log4net, in favor
of raising a Log event with appropriate arguments that can then be
passed to log4net (or any other logging framework). This is to make it
easier to consume without dependencies on other packages.